requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
openai==1.57.0
python-dotenv==1.0.1
httpx==0.27.2
